audio: fix play_pipeline trace

+++ audio/common/play_pipeline.c
@@ -201,8 +201,8 @@ void *play_pipeline_init(void *parameters)
 		codec_set_format(cid, mclk, rate, ctx->bit_width);
 	}
 
-	log_info("Starting pipeline (Sample Rate: %d Hz, Period: %zd frames)\n",
-			rate, period);
+	log_info("Starting pipeline (Sample Rate: %d Hz, Period: %u frames)\n",
+			rate, (uint32_t)period);
 
 	return ctx;
